Create String
Input a string from the user. Create a new string called ‘result’ in which you will replace the letter ‘e’ in the original string with letter ‘i’. 
Example : 
original = “eabcdef’ ; result = “iabcdif”
Original = “xyz” ; result = “xyz”

In this file i have create a code for the understanding of String in java through the given problem in which we will see how to declare a String, intitalization of a string and how to take input from the user of type String.