Create ARRAY_BASIC-16
Given an array arr of size n which contains elements in range from 0 to n-1, you need to find all the elements occurring more than once in the given array. Return the answer in ascending order. If no such element is found, return list containing [-1]. 

Note: Try and perform all operations within the provided array. The extra (non-constant) ) space needs to be used only for the array to be returned.