Create LinkedList_13
You are given two non-negative numbers 'num1' and 'num2' represented in the form of linked lists.



The digits in the linked lists are stored in reverse order, i.e. starting from least significant digit (LSD) to the most significant digit (MSD), and each of their nodes contains a single digit.



Calculate the sum of the two numbers and return the head of the sum list.